pyproject.toml
[project]
name = "06-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.4.3",
    "einops>=0.8.1",
    "faiss-cpu",
    "ipykernel>=7.1.0",
    "numpy<2",
    "openai>=2.15.0",
    "pandas>=2.3.3",
    "sentence-transformers>=5.2.0",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.5",
]
